requirements.txt
# --- parsing ---
pymupdf==1.24.9
python-pptx==1.0.2
pandas==2.2.2
pillow==10.4.0

# --- OCR ---
pytesseract==0.3.13

# --- embeddings (text + image) ---
sentence-transformers==3.0.1
torch>=2.3.0,<2.6
open-clip-torch==2.26.1

# --- vector & lexical search clients ---
qdrant-client==1.9.2
meilisearch==0.28.3

# --- retrieval helpers ---
bm25s==0.2.1
datasketch==1.6.5
pydantic==2.9.2
pyyaml==6.0.1

# --- api (optional) ---
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
python-dotenv==1.0.1

# --- utils / eval ---
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.4
